账单计算修正
账单计算修正

+++ src/com/linestore/dao/BillDao.java
@@ -10,13 +10,13 @@ import com.linestore.vo.Bill;
 public interface BillDao {
 	public List<Bill> select(Page page,Integer id);
 	public List<Bill> selectAll(Page page);
-	public List<Bill> selectAllByTime(Page page,String timeMin, String timeMax,Float amountMin, Float amountMax);
+	public List<Bill> selectAllByTime(Page page,String timeMin, String timeMax,BigDecimal amountMin, BigDecimal amountMax);
 	public List<Bill> search(String keywords);
 	public int queryAll(String hql);
 	public String mkData();
 	public void addBill(Bill bill);
 	public Bill selectById(Integer id);
-	public List<Bill> selectByTime(Page page,Integer id,String timeMin, String timeMax,Float amountMin, Float amountMax);
+	public List<Bill> selectByTime(Page page,Integer id,String timeMin, String timeMax,BigDecimal amountMin, BigDecimal amountMax);
 	
 	public List<Bill> queryByCusId(int cusId);
