+++ src/com/linestore/service/BillService.java
@@ -11,7 +11,7 @@ public interface BillService {
 	public List<Bill> select(Page page, Integer id);
 
 	public List<Bill> selectAll(Page page);
-	public List<Bill> selectAllByTime(Page page,String timeMin, String timeMax,Float amountMin, Float amountMax);
+	public List<Bill> selectAllByTime(Page page,String timeMin, String timeMax,BigDecimal amountMin, BigDecimal amountMax);
 	public List<Bill> search(String keywords);
 
 	public int queryAll(String hql);
@@ -21,7 +21,7 @@ public interface BillService {
 	public void addBill(Bill bill);
 
 	public Bill selectById(Integer id);
-	public List<Bill> selectByTime(Page page,Integer id,String timeMin, String timeMax,Float amountMin, Float amountMax);
+	public List<Bill> selectByTime(Page page,Integer id,String timeMin, String timeMax,BigDecimal amountMin, BigDecimal amountMax);
 	
 	public List<Bill> queryByCusId(int cusId);
